Add the vcl command.
Opens a file in a separate gvim window and frees the shell. The gvim window
is tied to the shell from which vcl was invoked. Subsequent invocations
of vcl will open files in that same gvim window.
If the filename has a ":<line_no>" suffix (like the ones you get on
compilation errors), the file will be opened with the cursor pointing to
that line.